requirements-pi.txt
absl-py==1.4.0
asn1crypto~=0.24.0
astroid==2.1.0
asttokens==1.1.13
attrs==23.1.0
automationhat==0.2.0
beautifulsoup4==4.7.1
blinker==1.4
blinkt==0.1.2
buttonshim==0.0.2
Cap1xxx==0.1.3
certifi==2018.8.24
chardet==3.0.4
Click~=7.0
colorama==0.3.7
colorzero==1.1
cookies==2.2.1
cryptography==2.6.1
cycler==0.11.0
docutils==0.14
drumhat==0.1.0
entrypoints==0.3
envirophat==1.0.0
exceptiongroup==1.1.1
ExplorerHAT==0.4.2
fonttools==4.38.0
fourletterphat==0.1.0
gpiozero==1.6.2
html5lib==1.0.1
idna==2.6
importlib-metadata==6.7.0
iniconfig==2.0.0
isort==4.3.4
itsdangerous==0.24
jedi==0.13.2
Jinja2==2.10
keyring~=17.1.1
keyrings.alt~=3.1.1
kiwisolver==1.4.4
lazy-object-proxy==1.3.1
logilab-common==1.4.2
lxml==4.3.2
MarkupSafe==1.1.0
matplotlib==3.5.3
mccabe==0.6.1
mediapipe-rpi3==0.8.8
mote==0.0.4
motephat==0.0.3
mypy==0.670
mypy-extensions==0.4.1
numpy==1.20.3
oauthlib==2.1.0
olefile==0.46
packaging==23.1
pantilthat==0.0.7
parso==0.3.1
pexpect==4.6.0
phatbeat==0.1.1
pianohat==0.1.0
piglow==1.2.5
pigpio==1.78
Pillow==9.5.0
pluggy==1.2.0
protobuf==4.23.3
psutil==5.5.1
pycairo==1.16.2
pycrypto==2.6.1
Pygments==2.3.1
pyinotify==0.9.6
PyJWT==1.7.0
pylint==2.2.2
pyOpenSSL==19.0.0
pyparsing==3.1.0
pyserial==3.4
pytest==7.4.0
python-dateutil==2.8.2
pyxdg==0.25
rainbowhat==0.1.0
reportlab==3.5.13
requests~=2.21.0
requests-oauthlib==1.0.0
responses==0.9.0
roman==2.0.0
RPi.GPIO==0.7.0
RTIMULib==7.2.1
scrollphat==0.0.7
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.16.0
six==1.12.0
skywriter==0.0.7
sn3218==1.2.7
soupsieve==1.8
spidev==3.5
ssh-import-id==5.7
thonny==3.3.10
tomli==2.0.1
touchphat==0.0.1
twython==3.7.0
typed-ast==1.3.1
typing-extensions==4.6.3
urllib3~=1.24.1
webencodings==0.5.1
Werkzeug==0.14.1
wrapt==1.10.11
zipp==3.15.0